vhost-device-media: move MediaArgs to module
Move MediaArgs and BackendType into a dedicated args module so
that it can be reused by the xtask crate to generate a manpage
in a followup commit.

The From<MediaArgs> for VuMediaConfig conversion is moved to
lib.rs where both types are defined.
