+++ vhost-device-media/src/main.rs
@@ -4,38 +4,8 @@
 
 extern crate vhost_device_media;
 
-use std::path::PathBuf;
-
 use clap::Parser;
-use vhost_device_media::{start_backend, BackendType, Error, VuMediaConfig};
-
-#[derive(Clone, Parser, Debug)]
-#[clap(author, version, about, long_about = None)]
-struct MediaArgs {
-    /// Unix socket to which a hypervisor connects to and sets up the control
-    /// path with the device.
-    #[clap(short, long)]
-    socket_path: PathBuf,
-
-    /// Path to the V4L2 media device file. Defaults to /dev/video0.
-    #[clap(short = 'd', long, default_value = "/dev/video0")]
-    v4l2_device: PathBuf,
-
-    /// Media backend to be used.
-    #[clap(short, long, default_value = "null")]
-    #[clap(value_enum)]
-    backend: BackendType,
-}
-
-impl From<MediaArgs> for VuMediaConfig {
-    fn from(args: MediaArgs) -> Self {
-        Self {
-            socket_path: args.socket_path,
-            v4l2_device: args.v4l2_device,
-            backend: args.backend,
-        }
-    }
-}
+use vhost_device_media::{args::MediaArgs, start_backend, Error, VuMediaConfig};
 
 fn main() -> std::result::Result<(), Error> {
     env_logger::init();
@@ -45,8 +15,11 @@ fn main() -> std::result::Result<(), Error> {
 
 #[cfg(test)]
 mod tests {
+    use std::path::PathBuf;
+
     #[cfg(any(feature = "simple-capture", feature = "v4l2-proxy", feature = "ffmpeg"))]
     use rstest::*;
+    use vhost_device_media::BackendType;
 
     use super::*;
